+++ src/mako/http/Response.php
@@ -13,7 +13,7 @@ use mako\http\response\Headers;
 use mako\http\response\senders\ResponseSenderInterface;
 use mako\http\response\Status;
 use mako\http\response\StatusInterface;
-use mako\security\Signer;
+use mako\security\signer\Signer;
 
 use function hash;
 use function header;