+++ src/mako/security/signer/exceptions/SignerException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\security\signer\exceptions;
+
+use mako\security\exceptions\SecurityException;
+
+/**
+ * Signer exception.
+ */
+class SignerException extends SecurityException
+{
+
+}