+++ tests/unit/security/signer/SignerTest.php
@@ -5,9 +5,10 @@
  * @license   http://www.makoframework.com/license
  */
 
-namespace mako\tests\unit\security;
+namespace mako\tests\unit\security\signer;
 
-use mako\security\Signer;
+use mako\security\signer\exceptions\SignerException;
+use mako\security\signer\Signer;
 use mako\tests\TestCase;
 use PHPUnit\Framework\Attributes\Group;
 
@@ -54,5 +55,34 @@ class SignerTest extends TestCase
 		$signer = new Signer('foobar');
 
 		$this->assertFalse($signer->validate(str_repeat('0', 64) . $string));
+	}
+
+	/**
+	 *
+	 */
+	public function testValidateOrThrowValid(): void
+	{
+		$string = 'hello, world!';
+
+		$signer = new Signer('foobar');
+
+		$signed = $signer->sign($string);
+
+		$this->assertEquals($string, $signer->validateOrThrow($signed));
+	}
+
+	/**
+	 *
+	 */
+	public function testValidateOrThrowInvalid(): void
+	{
+		$this->expectException(SignerException::class);
+		$this->expectExceptionMessageIs('Failed to validate the signed string. The signature is invalid or the data has been tampered with.');
+
+		$string = 'hello, world!';
+
+		$signer = new Signer('foobar');
+
+		$signer->validateOrThrow(str_repeat('0', 64) . $string);
 	}
 }